requirements.txt
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.9.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
deepgram-sdk==3.3.1
distro==1.9.0
dnspython==2.6.1
elevenlabs==1.13.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.4.1
groq==0.7.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
jsonpath-python==1.0.6
jupyter_client==8.6.2
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.56.0
orjson==3.10.5
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
retell==0.5.1
retell-sdk==4.8.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
twilio==9.3.7
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
yarl==1.9.4
